backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0
mysql-connector-python>=8.2.0
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.4.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
google-generativeai>=0.3.2
joblib>=1.3.2
Werkzeug>=3.0.1